+++ pokeRest/adapters/orm/gormio/schema/battle_opponent_party.go
@@ -1,13 +1,13 @@
 package schema
 
 type BattleOpponentParty struct {
-	ID                 uint  `gorm:"primaryKey;autoIncrement:true"`
-	OpponentPokemonId1 *uint // M:1 <- Pokemon
-	OpponentPokemonId2 *uint
-	OpponentPokemonId3 *uint
-	OpponentPokemonId4 *uint
-	OpponentPokemonId5 *uint
-	OpponentPokemonId6 *uint
+	ID                 uint64  `gorm:"primaryKey;autoIncrement:true"`
+	OpponentPokemonId1 *uint16 // M:1 <- Pokemon
+	OpponentPokemonId2 *uint16
+	OpponentPokemonId3 *uint16
+	OpponentPokemonId4 *uint16
+	OpponentPokemonId5 *uint16
+	OpponentPokemonId6 *uint16
 }
 
 func (BattleOpponentParty) TableName() string {
