+++ pokeRest/domain/value/pokemon_base_stats.go
@@ -1,18 +1,18 @@
 package value
 
 type PokemonBaseStats struct {
-	value int
+	value uint8
 }
 
-func NewPokemonBaseStats(value int) PokemonBaseStats {
+func NewPokemonBaseStats(value uint64) PokemonBaseStats {
 	if value < 0 {
 		return PokemonBaseStats{0}
 	} else if value > 252 {
 		return PokemonBaseStats{252}
 	}
-	return PokemonBaseStats{value}
+	return PokemonBaseStats{uint8(value)}
 }
 
-func (s PokemonBaseStats) Value() int {
+func (s PokemonBaseStats) Value() uint8 {
 	return s.value
 }
